% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmenters.R
\name{augment_facility_relocated}
\alias{augment_facility_relocated}
\title{Find distance from relocated and proposed new sites}
\usage{
augment_facility_relocated(proposed_facility, existing_facility)
}
\arguments{
\item{proposed_facility}{facilities proposed for the model - but this data
has extra information (\code{is_installed}) in it.}

\item{existing_facility}{facilities existing for the model - but this data
has extra information (\code{is_relocated}) in it.}
}
\value{
dataframe
}
\description{
This takes the proposed sites and the existing sites, with additional
information from the model, and then returns a dataframe of all of the
existing facilities that were relocated, and provides the distance to the
nearest facility, which is presumably the location to which it was
relocated to.
}
\examples{

\dontrun{

mc_cv_n100_test |>
  mutate(facility_distances = map2(
    .x = proposed_facility,
    .y = existing_facility,
    .f = augment_facility_relocated)) |>
  select(facility_distances) |>
  .[[1]]

}

}
