% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{augment_user}
\alias{augment_user}
\title{Augment the users data; add useful information}
\usage{
augment_user(facilities_selected, existing_facilities, existing_users)
}
\arguments{
\item{facilities_selected}{dataframe of facilities selected, obtained from
\code{extract_facility_selected}}

\item{existing_facilities}{existing facilities}

\item{existing_users}{existing users}
}
\value{
tibble of users, with distances between each user and facility
}
\description{
This returns the \code{user} dataframe, with added columns containing distance
between that user and a given facility - IDs are generated for IDs and
facilities that correspond to their row number.
}
\examples{

\dontrun{

mc_facilities_selected <-extract_facility_selected(
  solution_vector = x$lp_solution$solution,
  A_mat = x$A,
  proposed_facilities = x$proposed_facility)

augmented_users <- augment_user(
    facilities_selected = mc_facilities_selected,
    existing_facilities = mc_cv_fit_n20_test_1$existing_facility,
    existing_users = mc_cv_fit_n20_test_1$existing_user
    )
}

}
