% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_distance_matrix}
\alias{binary_distance_matrix}
\title{(Internal) Create a binary distance matrix}
\usage{
binary_distance_matrix(facility, user, distance_cutoff, d_proposed_user = NULL)
}
\arguments{
\item{facility}{data.frame of facilities}

\item{user}{data.frame of users}

\item{distance_cutoff}{integer of distance to use for cutoff}

\item{d_proposed_user}{Option distance matrix between proposed facilities and
users (see Examples).}
}
\value{
a logical matrix, of 1 if distance
between element i, j is less than or equal to the distance_cutoff, and
0 otherwise.
}
\description{
This is a wrapper function that returns a logical matrix, of 1 if distance
between element i, j is less than or equal to the distance_cutoff, and
0 otherwise.
}
