% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary_matrix_cpp}
\alias{binary_matrix_cpp}
\title{Create a binary matrix TRUE if distance satisfies a condition}
\usage{
binary_matrix_cpp(facility, user, distance_cutoff)
}
\arguments{
\item{facility}{a matrix with longitude and latitude in the first two columns}

\item{user}{a matrix with longitude and latitude in the first two columns}

\item{distance_cutoff}{numeric indicating the distance cutoff (in metres)
you are interested in. If a number is less than distance_cutoff, it will be
1, if it is greater than it, it will be 0.}
}
\value{
a logical matrix 1 if distance between element\link{i,j} is less than or
equal to the distance_cutoff, and 0 otherwise
}
\description{
Create a binary matrix TRUE if distance satisfies a condition
}
