% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/spherical_distance.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)

deg2rad(deg)
}
\arguments{
\item{deg}{degrees}

\item{deg}{A numeric vector in units of degrees.}
}
\value{
radians

The input numeric vector, converted to units of radians.
}
\description{
Convert degrees to radians

Convert from degrees to radians
}

