% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_matrix_cpp}
\alias{distance_matrix_cpp}
\title{Create a matrix of distances between two areas}
\usage{
distance_matrix_cpp(facility, user)
}
\arguments{
\item{facility}{a matrix with longitude and latitude in the first two columns}

\item{user}{a matrix with longitude and latitude in the first two columns}
}
\value{
a matrix of distances in metres between each user and facility,
with nrow(user) rows and nrow(facility) columns.
}
\description{
Create a matrix of distances between two areas
}

