% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_existing_coverage}
\alias{extract_existing_coverage}
\title{Extract the existing coverage}
\usage{
extract_existing_coverage(existing_facilities, existing_users,
  distance_cutoff)
}
\arguments{
\item{existing_facilities}{the existing facilities}

\item{existing_users}{the existing users}

\item{distance_cutoff}{the distance cutoffs}
}
\value{
tibble of existing coverage
}
\description{
Extract the existing coverage
}
\examples{

\dontrun{
extract_existing_coverage(existing_facilities = x$existing_facility,
   existing_users = x$existing_user,
   distance_cutoff = x$distance_cutoff)
   }
}
