% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_facility_selected}
\alias{extract_facility_selected}
\title{Extract Selected Facilities}
\usage{
extract_facility_selected(solution_vector, A_mat, proposed_facilities)
}
\arguments{
\item{solution_vector}{vector from lp_solution$solution}

\item{A_mat}{The "A" matrix from the solver}

\item{proposed_facilities}{Dataframe of proposed facilities}
}
\value{
dataframe of selected facilities
}
\description{
This takes the linear programming solution, the A matrix, and the proposed
facilities. It returns a tibble, which contains the facilities chosen from
the proposed facilities.
}
\examples{

# assuming that you've run max_coverage using lpSolve, then you
# will save the model output before the extraction process
# as `x`.
\dontrun{
mc_facilities_selected <- extract_facility_selected(
  solution_vector = x$lp_solution$solution,
  A_mat = x$A,
  proposed_facilities = x$proposed_facility)
  }

}
