% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_model_coverage}
\alias{extract_model_coverage}
\title{Extract a one-row summary of the model coverage}
\usage{
extract_model_coverage(augmented_user, distance_cutoff, n_added)
}
\arguments{
\item{augmented_user}{dataframe obtained from augment_user()}

\item{distance_cutoff}{numeric of the distance cutoff}

\item{n_added}{numeric of the number of facilities added}
}
\value{
tibble of summary coverage info
}
\description{
This function takes the users information, the distance cutoff, and the number of facilities added, and then returns a one-row dataframe containing summary information about the coverage.
}
\examples{

\dontrun{

augmented_users <- augment_user(
    facilities_selected = mc_facilities_selected,
    existing_facilities = x$existing_facility,
    existing_users = x$existing_user)

extract_model_coverage(
    augmented_user = augmented_users,
    distance_cutoff = x$distance_cutoff,
    n_added = x$n_added)
}

}
