% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_users_affected}
\alias{extract_users_affected}
\title{Extract users affected}
\usage{
extract_users_affected(A_mat, solution_vector, user_id, users_not_covered)
}
\arguments{
\item{A_mat}{A matrix}

\item{solution_vector}{The vector of solutions}

\item{user_id}{The IDs of the individuals}

\item{users_not_covered}{# those users not covered by original AEDs}
}
\value{
tibble taken from `users`, those who are affectd by new placements
}
\description{
Extract additional users affected by new coverage from the new facilities
}
\examples{

\dontrun{
extract_users_affected(
    A_mat = x$A,
    solution_vector = x$lp_solution$solution,
    user_id = x$user_id,
    users_not_covered = x$user_not_covered)
    }

}
