% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_to_user_distance.R
\name{facility_to_user_distance}
\alias{facility_to_user_distance}
\title{facility_to_user_distance}
\usage{
facility_to_user_distance(df_facility_lat, df_facility_long, df_user_lat,
  df_user_long)
}
\arguments{
\item{df_facility_lat}{a vector from a data frame, x, containing latitude}

\item{df_facility_long}{a vector from a dataframe, x, containing longitude}

\item{df_user_lat}{a vector from a data frame, y, containing latitude}

\item{df_user_long}{a vector from a data frame, y, containing longitude}
}
\value{
a list where each element contains all of the distances from of every Y to every X.
}
\description{
uses haversines formula to calculate the distance between two sets of points. It calculates the distance from every X to every y. This came form an example of finding the distance from AEDs (here listed as X) to each OHCA (here listed as Y)
}

