% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_user_dist.R
\name{facility_user_dist}
\alias{facility_user_dist}
\title{facility_user_dist}
\usage{
facility_user_dist(facility, user, coverage_distance = 100,
  nearest = "facility")
}
\arguments{
\item{facility}{a dataframe containing columns aed_id, lat, and long}

\item{user}{a dataframe containing columns ohca_id, lat, and long}

\item{coverage_distance}{numeric indicating the coverage level for the facilities to be within in metres to a user. Default value is 100 metres.}

\item{nearest}{character When "facility", it returns the dataframe with the nearest facilities to each user. When "user", it returns the dataframe with the nearest users to each facility. when NULL it returns the complete pairwise distances}
}
\value{
a data frame containing the distance between each aed and each ohca, with columns named aed_id, lat_aed, long_aed, ohca_id, lat_ohca, long_ohca, and distance - the distance in meters between each aed and ohca in a row.
}
\description{
Uses haversines formula to calculate the distance between lat/long co-ordinates of every facility and every user, returning a data_frame. You can think of "facilities" as something like mobile towers, police centres, or AED locations, and "users" as something like individual houses, crime locations, or heart attack locations. The motivating example for this function was finding the distance from Automatic Electronic Defibrillators (AEDs) to each Out of Hospital Cardiac Arrest (OHCA), where the locations for AEDs and OHCAs are in separate dataframes. Currently facifacility_user_dist makes the strict assumption that the facility and user dataframes have columns named aed_id, lat, and long, and ohca_id, lat, and long. This will be updated soon.
}

