% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_user_indic.R
\name{facility_user_indic}
\alias{facility_user_indic}
\title{facility_user_indic}
\usage{
facility_user_indic(facility, user, dist_indic)
}
\arguments{
\item{facility}{a dataframe containing columns aed_id, lat, long}

\item{user}{a dataframe containing columns ohca_id, lat, long}

\item{dist_indic}{the distance in meters that you are interested in}
}
\value{
a dataframe with variables ohca_id, and aed_id_number, with the id from each aed_id being transposed into each column name.
}
\description{
This is a data manipulation function for facility_user_dist. This function creates a spread matrix of the distances between each ohca and each aed. There is an ohca_id column, and then a column for each aed_id, with a given cell being the distance between an ohca in a row, and that column. This distance is converted into an indicator variable, based upon whether that distance is less than the provided dist_indic parameter. In the future I might change the dist_indic function to be optional, but this whole function mainly exists to make it easier to do the computation in the max_coverage function.
}

