% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.maxcovr}
\alias{is.maxcovr}
\title{Test if the object is a maxcovr object}
\usage{
is.maxcovr(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `maxcovr` class.
}
\description{
Test if the object is a maxcovr object
}
