% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.maxcovr_relocation}
\alias{is.maxcovr_relocation}
\title{Test if the object is a maxcovr_relocation object}
\usage{
is.maxcovr_relocation(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{maxcovr_relocation} class.
}
\description{
Test if the object is a maxcovr_relocation object
}
