% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_cov_relocation.R
\name{max_coverage_relocation}
\alias{max_coverage_relocation}
\title{Maximum Coverage when considering relocation}
\usage{
max_coverage_relocation(existing_facility = NULL, proposed_facility, user,
  distance_cutoff, n_added, n_solutions = 1, cost_install, cost_relocate,
  cost_total, solver = "lpSolve")
}
\arguments{
\item{existing_facility}{data.frame containing the facilities that are already in existing, with columns names lat, and long.}

\item{proposed_facility}{data.frame containing the facilities that are being proposed, with column names lat, and long.}

\item{user}{data.frame containing the users of the facilities, along with column names lat, and long.}

\item{distance_cutoff}{numeric indicating the distance cutoff (in metres)
you are interested in. If a number is less than distance_cutoff, it will be
1, if it is greater than it, it will be 0.}

\item{n_added}{the maximum number of facilities to add.}

\item{n_solutions}{Number of possible solutions to return. Default is 1.}

\item{cost_install}{integer the cost of installing a new facility}

\item{cost_relocate}{integer the cost of relocating a new facility}

\item{cost_total}{integer the total cost allocated to the project}

\item{solver}{character default is lpSolve, but currently in development is a Gurobi solver, see issue #25 : \url{https://github.com/njtierney/maxcovr/issues/25}}
}
\value{
dataframe of results
}
\description{
This function adds a relocation step
}
