% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcovr-package.R
\docType{package}
\name{maxcovr-package}
\alias{maxcovr-package}
\title{maxcovr: A Set of Tools For Solving The Maximal Covering Location Problem}
\description{
Solving the "maximal covering location problem" as described by Church can be difficult for users not familiar with linear programming. maxcovr provides functions to make it easy to solve this problem, and tools to calculate facility coverage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njtierney/maxcovr}
  \item \url{https://njtierney.github.io/maxcovr/}
  \item Report bugs at \url{https://github.com/njtierney/maxcovr/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com}
}

}
\keyword{internal}
