% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmenters.R
\name{n_installed}
\alias{n_installed}
\title{Extract the number of facilities installed}
\usage{
n_installed(proposed_facility)
}
\arguments{
\item{proposed_facility}{dataframe from the mc_model, of facilities proposed
with the additional information about whether the facility was installed
or not - \code{is_installed}}
}
\value{
datafrmae
}
\description{
Using the model-modified dataframe of \code{proposed_facility}, count the
number of events installed.
}
\examples{

\dontrun{

mc_cv_n100_test |>
    mutate(n_installed = map(
        .x = proposed_facility,
        .f = n_installed
    )) |>
    select(n_installed) |>
    .[[1]]

}

}
