% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmenters.R
\name{n_relocated}
\alias{n_relocated}
\title{Extract the number of facilities relocated.}
\usage{
n_relocated(existing_facility)
}
\arguments{
\item{existing_facility}{the facilities originally existing - this will have
to be output from the model, however.}
}
\value{
dataframe containing a single column of
}
\description{
Extract the number of facilities relocated.
}
\examples{

\dontrun{

mc_cv_n100_test \%>\%
  mutate(n_relocated = map(
    .x = existing_facility,
    .f = n_relocated)) \%>\%
  select(n_relocated) \%>\%
  .[[1]]

}

}
