% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearest_facility_dist}
\alias{nearest_facility_dist}
\title{nearest facility + distance to a user}
\usage{
nearest_facility_dist(facility, user)
}
\arguments{
\item{facility}{a matrix with longitude and latitude in the first two columns}

\item{user}{a matrix with longitude and latitude in the first two columns}
}
\value{
matrix with 3 columns: user_id, facility_id, distance, where the
user_id is the identifier for the user, the facility_id is the identifier
for the facility that is closest to that user, and the distance is the
distance in metres from that user to that facility.
}
\description{
nearest facility + distance to a user
}

