% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spherical_distance_cpp}
\alias{spherical_distance_cpp}
\title{Calculate distance using haversines formula}
\usage{
spherical_distance_cpp(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1}{latitude from the first location}

\item{long1}{longitude from the first location}

\item{lat2}{latitude from the second location}

\item{long2}{longitude from the second location}
}
\value{
distance in metres between two locations
}
\description{
Calculate distance using haversines formula
}
