% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmenters.R
\name{summarise_relocated_dist}
\alias{summarise_relocated_dist}
\title{Find the average distance from facilities relocated to their final place}
\usage{
summarise_relocated_dist(augment_facility_relocated)
}
\arguments{
\item{augment_facility_relocated}{dataframe from function:
\code{augment_facility_relocated}}
}
\value{
dataframe
}
\description{
This takes data from the function \code{augment_facility_relocated} function of
the same name and then summarises it to find the average and sd of the
distance between the two.
}
\examples{

\dontrun{

mc_cv_n100_test \%>\%
    mutate(
        facility_distances = map2(
            .x = proposed_facility,
            .y = existing_facility,
            .f = augment_facility_relocated
        ),
        summary_relocated_dist = map(
            .x = facility_distances,
            .f = summarise_relocated_dist
        )
    ) \%>\%
    # select(facility_distances) \%>\%
    select(summary_relocated_dist) \%>\%
    .[[1]]

}

}
