% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary_mc_cv}
\alias{summary_mc_cv}
\title{Summary for max_coverage cross validation}
\usage{
summary_mc_cv(model, test_data)
}
\arguments{
\item{model}{the cross validated model}

\item{test_data}{the cross validated test data}
}
\value{
a summary dataframe
}
\description{
Summary for max_coverage cross validation
}
\examples{

\dontrun{

library(maxcovr)
library(tidyverse)

york_selected <- york \%>\% filter(grade == "I")
york_unselected <- york \%>\% filter(grade != "I")

mc_cv <- modelr::crossv_kfold(york_crime, 5) \%>\%
                 mutate(test = map(test,as_tibble),
                 train = map(train,as_tibble))

mc_cv_fit <- map_df(mc_cv$train,
                    ~max_coverage(existing_facility = york_selected,
                    proposed_facility = york_unselected,
                    user = .,
                    n_added = 20,
                    distance_cutoff = 100))

 summary_mc_cv(mc_cv_fit,
               mc_cv$test)

}

}
