% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{summary_mc_cv_relocate}
\alias{summary_mc_cv_relocate}
\title{Summary for max_coverage cross validation for relocation models}
\usage{
summary_mc_cv_relocate(model, test_set)
}
\arguments{
\item{model}{the cross validated model}

\item{test_data}{the cross validated test data}
}
\value{
a summary dataframe
}
\description{
Summary for max_coverage cross validation for relocation models
}
\examples{

\dontrun{

library(maxcovr)
library(tidyverse)

york_selected <- york \%>\% filter(grade == "I")
york_unselected <- york \%>\% filter(grade != "I")

mc_cv <- modelr::crossv_kfold(york_crime, 5) \%>\%
                 mutate(test = map(test,as_tibble),
                 train = map(train,as_tibble))

mc_cv_relocate <- map_df(mc_cv$train,
                    ~max_coverage_relocation(existing_facility = york_selected,
                               proposed_facility = york_unselected,
                               user = .,
                               cost_install = 2500,
                               cost_removal = 700,
                               cost_total = 50000,
                               distance_cutoff = 100,
                               solver = "gurobi"))
#'
 summary_mc_cv_relocate(mc_cv_relocate, mc_cv$test)

}

}
