% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_to_y_distance.R
\name{x_to_y_distance}
\alias{x_to_y_distance}
\title{x_to_y_distance}
\usage{
x_to_y_distance(df_x_lat, df_x_long, df_y_lat, df_y_long)
}
\arguments{
\item{df_x_lat}{a vector from a data frame, x, containing latitude}

\item{df_x_long}{a vector from a dataframe, x, containing longitude}

\item{df_y_lat}{a vector from a data frame, y, containing latitude}

\item{df_y_long}{a vector from a data frame, y, containing longitude}
}
\value{
a list where each element contains all of the distances from of every Y to every X.
}
\description{
uses haversines formula to calculate the distance between two sets of points. It calculates the distance from every X to every y. This came form an example of finding the distance from AEDs (here listed as X) to each OHCA (here listed as Y)
}

