% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mc_result_relocation.R
\name{extract_mc_results_relocation}
\alias{extract_mc_results_relocation}
\title{(Internal) Summarise maxcovr relocation model with facility and user info}
\usage{
extract_mc_results_relocation(x)
}
\arguments{
\item{x}{the fitted model from max_coverage_relocation}
}
\value{
a list containing multiple dataframes summarising the model
}
\description{
\code{extract_mc_results_relocation} takes a fitted max_coverage object and
returns useful summary information from the model, specifically for the
relocation method.
}
