% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_user_dist.R
\name{facility_user_dist}
\alias{facility_user_dist}
\title{facility_user_dist}
\usage{
facility_user_dist(facility, user)
}
\arguments{
\item{facility}{a dataframe containing lat/long}

\item{user}{a dataframe containing}
}
\value{
a data frame containing the distance between each aed and each ohca
}
\description{
This is a more sensible implementation of facility_to_user_dist.
Basically it returns a dataframe, not a mega-uber-list.
I need to work on what you need to pass the function.
Currently facility needs to have columns called aed_id, lat, and long, and user needs to have columns called ohca_id, lat, and long.
}

