% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_user_indic.R
\name{facility_user_indic}
\alias{facility_user_indic}
\title{facility_user_indic}
\usage{
facility_user_indic(facility, user, indic = 0.1)
}
\arguments{
\item{facility}{a dataframe containing columns aed_id, lat, long}

\item{user}{a dataframe containing columns ohca_id, lat, long}

\item{indic}{the distance in km that you are interested in}
}
\value{
a dataframe with a column of ohca IDs
}
\description{
Return a spread distance matrix
}

