% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_users_not_covered}
\alias{find_users_not_covered}
\title{(Internal) Create a dataframe of the users not covered}
\usage{
find_users_not_covered(
  existing_facility,
  user,
  distance_cutoff,
  d_existing_user = NULL
)
}
\arguments{
\item{existing_facility}{data.frame of existing facilities}

\item{user}{data.frame of existing users}

\item{distance_cutoff}{integer of distance cutoff}

\item{d_existing_user}{Optional distance matrix between existing facilities
and users.}
}
\value{
data.frame of those users not covered by current facilities
}
\description{
(Internal) Create a dataframe of the users not covered
}
