% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_coverage.R
\name{max_coverage}
\alias{max_coverage}
\title{max_coverage}
\usage{
max_coverage(A, facility, user, n_added, n_solutions = 1,
  solver = "lpSolve")
}
\arguments{
\item{A}{is a spread data matrix for all of the distances, it is obtained using facility_user_dist, then facilit_user_indic.}

\item{facility}{data.frame containing an ohca_id, lat, and long}

\item{user}{data.frame containing an aed_id, lat, and long}

\item{n_added}{the maximum number of facilities to add.}

\item{n_solutions}{is the number of possible solutions to be returned. Default value is set to 1.}

\item{solver}{character default is lpSolve, but currently in development is a Gurobi solver}
}
\value{
returns
}
\description{
max_coverage solves the binary optimisation problem known as the "maximal covering location problem" as described by Church (http://www.geog.ucsb.edu/~forest/G294download/MAX_COVER_RLC_CSR.pdf). This package was implemented to make it easier to solve this problem in the context of the research initially presented by Chan et al (http://circ.ahajournals.org/content/127/17/1801.short) to identify ideal locations to place AEDs.
}

