% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_coverage.R
\name{max_coverage}
\alias{max_coverage}
\title{max_coverage}
\usage{
max_coverage(facility, user, dist_indic, num_aed, n_solutions = 1)
}
\arguments{
\item{facility}{data.frame containing an ohca_id, lat, and long}

\item{user}{data.frame containing an aed_id, lat, and long}

\item{dist_indic}{numeric stating the distance in kilometers for facility-user distance}

\item{num_aed}{is the maximum number of AEDs that can be found}

\item{n_solutions}{is the number of possible solutions to be returned. Default value is set to 1.}
}
\value{
returns
}
\description{
max_coverage solves the binary optimisation problem known as the "maximum location coverage problem" as described by Church.... This package was implemented to make it easier to solve this problem in the context of the research initially presented by Chan et al to identify ideal locations to place AEDs.
}

