% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_coverage.R
\name{summarise_coverage}
\alias{summarise_coverage}
\title{summarise_coverage}
\usage{
summarise_coverage(df_dist, dist_indic = 100, spread = TRUE)
}
\arguments{
\item{df_dist}{distance matrix, as computed by facility_user_dist}

\item{dist_indic}{the critical distance range that you would like to know,
default is 100m}

\item{spread}{do you want this in a one row summary, or a two row summary?
Default is TRUE, a one row summary}
}
\value{
dataframe
}
\description{
Provides summary information of the coverage, using the distance dataframe created by maxcovr::facility_user_dist().
}

